:art: style: accounts 앱에서 views 파일 생성 후 코드 분리 - account.py 생성, 계정 관리 - profile.py 생성, 프로필 관리
